+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.0.1
+## 0.0.3
+- Added a TimeIterable class.
+- Added a Month class.
 
+## 0.0.2
+
+## 0.0.1
 - Initial version